backend/requirements.txt
flask
flask-cors
networkx
matplotlib
python-dotenv
groq